requirements.txt
Flask==0.12.2
gunicorn==19.7.1
google-cloud-language
google-cloud-bigquery
tensorflow
tensorflow-serving-api
numpy
